# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/products/search": {
      "post": {
        "summary": "Search products",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products list"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "Get admin users",
        "responses": {
          "200": {
            "description": "List users"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Alternate customer login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "validLogin": {
                  "value": {
                    "username": "juan",
                    "password": "123456"
                  }
                },
                "adminLogin": {
                  "value": {
                    "username": "admin",
                    "password": "admin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login result"
          }
        }
      }
    },
    "/account/recovery/lookup": {
      "post": {
        "summary": "Account recovery lookup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "username": "juan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account lookup result"
          }
        }
      }
    },
    "/reports/products": {
      "post": {
        "summary": "Internal product inventory report",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "category": "computers",
                "featured": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product matches"
          },
          "500": {
            "description": "Report generation error"
          }
        }
      }
    },
    "/catalog/search": {
      "post": {
        "summary": "Public catalog search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "category": "computers",
                "maxPrice": 1000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product matches"
          },
          "400": {
            "description": "Generic invalid query error"
          }
        }
      }
    },
    "/catalog/advanced-search": {
      "post": {
        "summary": "Advanced catalog search for inventory analysts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "category": "computers",
                "maxPrice": 1000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products and query metadata"
          }
        }
      }
    },
    "/security/request-risk": {
      "post": {
        "summary": "Request risk evaluation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "path": "/orders/search",
                "filters": {
                  "status": "paid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suspicion flag and elapsed time"
          }
        }
      }
    },
    "/security/filter-review": {
      "post": {
        "summary": "Review dynamic filters submitted by integrations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "source": "partner-api",
                "filters": {
                  "category": "computers",
                  "featured": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risk assessment"
          }
        }
      }
    },
    "/orders/search": {
      "post": {
        "summary": "Support order search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "status": "paid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order matches"
          }
        }
      }
    },
    "/admin/staff-directory": {
      "get": {
        "summary": "Legacy staff directory",
        "responses": {
          "200": {
            "description": "List privileged users"
          }
        }
      }
    },
    "/admin/staff": {
      "get": {
        "summary": "Protected staff directory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List privileged users"
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Missing token or insufficient permissions"
          }
        }
      }
    },
    "/products/catalog": {
      "post": {
        "summary": "Frontend product catalog search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "category": "computers",
                "maxPrice": 1000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product matches"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}